+++ src/ffmpegio/transcode.py
@@ -1,5 +1,4 @@
-from collections.abc import Sequence
-from . import ffmpegprocess, configure, utils, FFmpegError
+from . import ffmpegprocess as fp, configure, utils, FFmpegError
 
 __all__ = ["transcode"]
 
@@ -13,13 +12,13 @@ def transcode(
     two_pass=False,
     pass1_omits=None,
     pass1_extras=None,
-    **options
+    **options,
 ):
     """Transcode media files to another format/encoding
 
     :param inputs: url/path of the input media file or a sequence of tuples, each
                    containing an input url and its options dict
-    :type inputs: str or sequence of (str,dict)
+    :type inputs: str or a list of str or a sequence of (str,dict)
     :param outputs: url/path of the output media file or a sequence of tuples, each
                     containing an output url and its options dict
     :type outputs: str or sequence of (str, dict)
@@ -34,15 +33,17 @@ def transcode(
     :type show_log: bool, optional
     :param two_pass: True to encode in 2-pass
     :param pass1_omits: list of output arguments to ignore in pass 1, defaults to
-                        None (removes 'c:a' or 'acodec')
-    :type pass1_omits: seq(str), optional
+                        None (removes 'c:a' or 'acodec'). For multiple outputs,
+                        specify use list of the list of arguments, matching the
+                        length of outputs, for per-output omission.
+    :type pass1_omits: seq(str), or seq(seq(str)) optional
     :param pass1_extras: list of additional output arguments to include in pass 1,
                          defaults to None (add 'an' if `pass1_omits` also None)
     :type pass1_extras: dict(int:dict(str)), optional
     :param \\**options: FFmpeg options. For output and global options, use FFmpeg
-                        option names as is. For input options, prepend "input\_" to
-                        the option name. For example, input_r=2000 to force the
-                        input frame rate to 2000 frames/s (see :doc:`options`).
+                        option names as is. For input options, append "_in" to the
+                        option name. For example, r_in=2000 to force the input frame
+                        rate to 2000 frames/s (see :doc:`options`).
 
                         If multiple inputs or outputs are specified, these input
                         or output options specified here are treated as common
@@ -50,8 +51,6 @@ def transcode(
                         ``inputs`` or ``outputs`` sequence will overwrite those
                         specified here.
     :type \\**options: dict, optional
-    :return: returncode of FFmpeg subprocess
-    :rtype: int
 
 
     """
@@ -60,22 +59,38 @@ def transcode(
     input_options = utils.pop_extra_options(options, "_in")
     global_options = utils.pop_global_options(options)
 
-    # detect single input/output argument
-    if isinstance(inputs, str) or not isinstance(inputs, Sequence):
-        inputs = [(inputs, None)]
-    if isinstance(outputs, str) or not isinstance(outputs, Sequence):
-        outputs = [(outputs, None)]
+    def format_arg(arg, defopts):
+        def test(a, is_list):
+            try:
+                assert len(a) == 2
+                assert isinstance(a[1], dict)
+                return (a[0], {**defopts, **a[1]})
+            except:
+                if is_list:
+                    return (a, defopts)
+                raise
+
+        # special case: a list of inputs w/out options
+        if type(arg) == list:
+            return [test(a, True) for a in arg]
+
+        # attempt to map url-options pairs
+        try:
+            return [test(a, False) for a in arg]
+        except:
+            return [(arg, defopts)]
+
+    inputs = format_arg(inputs, input_options)
+    outputs = format_arg(outputs, options)
 
     # initialize FFmpeg argument dict
     args = configure.empty(global_options)
 
     for url, opts in inputs:
-        opts = {**input_options, **(opts or {})}
         input_url, stdin, input = configure.check_url(url, False, opts.get("f", None))
         configure.add_url(args, "input", input_url, opts)
 
     for url, opts in outputs:
-        opts = {**options, **(opts or {})}
         output_url, stdout, _ = configure.check_url(url, True)
         i, _ = configure.add_url(args, "output", output_url, opts)
 
@@ -83,15 +98,10 @@ def transcode(
         configure.build_basic_vf(args, None, i)
 
     kwargs = (
-        {
-            "pass1_omits": None if pass1_omits is None else [pass1_omits],
-            "pass1_extras": None if pass1_extras is None else [pass1_extras],
-        }
-        if two_pass
-        else {}
+        {"pass1_omits": pass1_omits, "pass1_extras": pass1_extras} if two_pass else {}
     )
 
-    pout = (ffmpegprocess.run_two_pass if two_pass else ffmpegprocess.run)(
+    pout = (fp.run_two_pass if two_pass else fp.run)(
         args,
         progress=progress,
         overwrite=overwrite,