+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Changed
+
+- `transcode()` allowed to take multiple inputs or outputs
+
 ## [0.5.0] - 2022-04-03
 
 ### Added